+++ research/conditional/utils/argparse.py
@@ -189,5 +189,7 @@ def introduce_parser_arguments(
     parser.add_argument("--blanks_add_embedding", action="store_true")
     parser.add_argument("--blanks_residual", action="store_true")
     parser.add_argument("--blanks_learnable_weights", action="store_true")
+    parser.add_argument("--blank_initial_weight", type=float, default=1.0)
+    parser.add_argument("--blanks_use_straight_through", action="store_true")
 
     return parser
