+++ research/conditional/utils/conditional_trainer.py
@@ -7,7 +7,7 @@ from typing import Callable, Optional, Literal
 
 import torch
 from attr import define
-from lizrd.core.llm import BlankDiffPredictionHead
+from research.blanks.model import BlankDiffPredictionHead
 from lizrd.core.misc import propagate_forward_pass_cache
 from lizrd.support.decoding import decode_single_example
 from lizrd.support.logging import AbstractLogger
